raspberry_pi/requirements.txt
# Raspberry Pi Dependencies for RoadGuard
# Install with: pip install -r requirements.txt

# ─── Core ────────────────────────
ultralytics>=8.0.0          # YOLOv8 for pothole detection
opencv-python>=4.8.0        # Camera capture and image processing
numpy>=1.24.0               # Numerical operations

# ─── Hardware Drivers ────────────
pyserial>=3.5               # GPS module (NEO-6M) serial communication
gpiozero>=2.0.0             # Modern GPIO control for buzzer (Bookworm optimized)

# ─── Networking ──────────────────
requests>=2.31.0            # HTTP client to send data to Flask API

# ─── Optional (for GPS parsing) ──
pynmea2>=1.19.0             # Advanced NMEA sentence parsing (optional)
roboflow